+++ packages/lumx-core/src/js/constants/index.ts
@@ -14,6 +14,7 @@ export * from './keycodes';
  */
 export const DIALOG_TRANSITION_DURATION = 400;
 export const NOTIFICATION_TRANSITION_DURATION = 200;
+export const SLIDESHOW_TRANSITION_DURATION = 5000;
 
 /**
  * Delay on hover after which we open or close the tooltip.